+++ sat/util.py
@@ -1,25 +1,28 @@
+#
+# MIT License
+#
+# (C) Copyright 2019-2022 Hewlett Packard Enterprise Development LP
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included
+# in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+# OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+# ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+# OTHER DEALINGS IN THE SOFTWARE.
+#
 """
 Contains structures and code that is generally useful across all of SAT.
-
-(C) Copyright 2019-2022 Hewlett Packard Enterprise Development LP.
-
-Permission is hereby granted, free of charge, to any person obtaining a
-copy of this software and associated documentation files (the "Software"),
-to deal in the Software without restriction, including without limitation
-the rights to use, copy, modify, merge, publish, distribute, sublicense,
-and/or sell copies of the Software, and to permit persons to whom the
-Software is furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included
-in all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL
-THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR
-OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
-ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
-OTHER DEALINGS IN THE SOFTWARE.
 """
 from json.encoder import JSONEncoder
 import sys
@@ -373,6 +376,31 @@ class SATEncoder(JSONEncoder):
 json_dump = partial(dumps, cls=SATEncoder, **JSON_FORMAT_PARAMS)
 
 
+def get_resource_section_path(section):
+    """Get the path to a section in the sat resource directory.
+
+    If the section subdirectory does not exist, it is created. If it cannot be
+    created, SystemExit is raised.
+
+    Args:
+        section (str): A subdirectory under the resource directory
+
+    Returns:
+        str: the path to the directory in the SAT resource directory
+
+    Raises:
+        SystemExit: if the given subdirectory does not exist and cannot be
+            created
+    """
+    resource_path = os.path.join(os.environ['HOME'], '.config', 'sat', section)
+    try:
+        os.makedirs(resource_path, exist_ok=True)
+    except OSError as err:
+        LOGGER.error("Unable to create resource directory '%s': %s", resource_path, err)
+        raise SystemExit(1)
+    return resource_path
+
+
 def get_resource_filename(name, section='.'):
     """Get the pathname to a resource file.
 
@@ -384,14 +412,7 @@ def get_resource_filename(name, section='.'):
         Full pathname to the resource file.
     """
 
-    resource_path = os.path.join(os.environ['HOME'], '.config', 'sat', section)
-
-    try:
-        os.makedirs(resource_path, exist_ok=True)
-    except OSError as err:
-        LOGGER.error("Unable to create resource directory '%s': %s", resource_path, err)
-        raise SystemExit(1)
-
+    resource_path = get_resource_section_path(section)
     return os.path.join(resource_path, name)
 
 
@@ -622,3 +643,32 @@ def match_query_key(query_key, headings):
                        f"Using first match: '{matching_keys[0]}' from {tuple(matching_keys)}.")
 
     return matching_keys[0]
+
+
+def ensure_permissions(path, file_mode=0o600, dir_mode=0o700):
+    """Lock down the file and directory permissions for a given path.
+
+    Specifically, if the given path points to a file, the file at the given
+    path will be given the mode `file_mode`, and the containing directory will
+    be given the mode `dir_mode`. If the given path points to a directory, it
+    will be given the mode `dir_mode`.
+
+    No action is taken for the given file or its containing directory if each
+    does not exist.
+
+    Args:
+        path (str): the path to the file or directory to have permissions set
+        file_mode (int): the mode of the file
+        dir_mode (int): the mode of the containing directory
+
+    Returns: None
+    """
+    if os.path.isdir(path):
+        os.chmod(path, dir_mode)
+        return
+    elif os.path.isfile(path):
+        os.chmod(path, file_mode)
+
+    dir_path = os.path.dirname(path)
+    if os.path.isdir(dir_path):
+        os.chmod(dir_path, dir_mode)